requirements.txt
urllib3==1.25.7
xmltodict==0.11.0
requests==2.22.0
requests-oauthlib==1.0.0
python-twitter==3.5
mysql==0.0.2
mysql-connector-python==8.0.13
chardet==3.0.4
